requirements.txt
accelerate==0.24.1
aiofiles==23.2.1
altair==5.1.2
annotated-types==0.6.0
anyio==3.7.1
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.1.1
cycler==0.12.1
diffusers==0.23.0
exceptiongroup==1.1.3
fastapi==0.104.1
ffmpy==0.3.1
filelock==3.13.1
fonttools==4.44.0
fsspec==2023.10.0
gradio==4.3.0
gradio_client==0.7.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.1
huggingface-hub==0.17.3
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.1
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
kiwisolver==1.4.5
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.3
mdurl==0.1.2
mpmath==1.3.0
networkx==3.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.52
nvidia-nvtx-cu12==12.1.105
orjson==3.9.10
packaging==23.2
pandas==2.0.3
Pillow==10.1.0
pkgutil_resolve_name==1.3.10
psutil==5.9.6
pydantic==2.5.0
pydantic_core==2.14.1
pydub==0.25.1
Pygments==2.16.1
pyparsing==3.1.1
python-dateutil==2.8.2
python-multipart==0.0.6
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.30.2
regex==2023.10.3
requests==2.31.0
rich==13.6.0
rpds-py==0.12.0
safetensors==0.4.0
scikit-learn==1.3.2
scipy==1.10.1
semantic-version==2.10.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
sympy==1.12
threadpoolctl==3.2.0
tokenizers==0.14.1
tomlkit==0.12.0
toolz==0.12.0
torch==2.1.1
torchvision==0.16.1
tqdm==4.66.1
transformers==4.35.0
triton==2.1.0
typer==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
uvicorn==0.24.0.post1
websockets==11.0.3
zipp==3.17.0
